feat(plc): make Home quick-create a top button bar; body is Attention + Recent Docs
Converts the Quick Create card into a horizontal button bar in the Home
header (under the title/members row). With members and quick-create now in
the header, the body is a clean two-column split: Attention (wider) and
Recent Docs. Replaces QuickCreateCard with QuickCreateBar.
