+++ components/plc/home/PlcHome.tsx
@@ -2,14 +2,14 @@
  * PlcHome — the redesigned landing page for a PLC Dashboard session.
  *
  * Replaces the legacy draggable bento/grid overview with a clean,
- * responsive two-column layout. NOT a draggable grid — just cards.
+ * responsive layout. NOT a draggable grid — just cards.
  *
  * Layout:
- *   - Members live in the page header (cluster of avatars, right-aligned).
- *   - lg+: two columns filling the full content width — a wider primary
- *          column (Attention + Recent Docs) and a narrower sidebar column
- *          (Quick Create). The two columns flow independently (no forced
- *          row alignment), so there are no empty grid holes.
+ *   - Header: title/subtitle (left) + members cluster (right), with a
+ *     quick-create button bar directly beneath.
+ *   - lg+: two columns filling the full content width — Attention (wider)
+ *          and Recent Docs (narrower). Columns flow independently, so there
+ *          are no empty grid holes.
  *   - sm/md: single column — cards stack in priority order.
  *
  * Design intent: "calm, clean, professional" — glassmorphism card surfaces,
@@ -22,7 +22,7 @@ import { useTranslation } from 'react-i18next';
 import type { Plc } from '@/types';
 import type { PlcSectionId } from '../sections';
 import { AttentionCard } from './cards/AttentionCard';
-import { QuickCreateCard } from './cards/QuickCreateCard';
+import { QuickCreateBar } from './cards/QuickCreateBar';
 import { RecentDocsCard } from './cards/RecentDocsCard';
 import { MembersHeaderCluster } from './cards/MembersHeaderCluster';
 
@@ -36,33 +36,37 @@ export const PlcHome: React.FC<PlcHomeProps> = ({ plc, onNavigate }) => {
 
   return (
     <div className="flex flex-col h-full overflow-y-auto">
-      {/* Page header — title/subtitle left, members cluster right */}
-      <div className="px-6 pt-6 pb-4 border-b border-slate-100 bg-white/50 backdrop-blur-sm flex items-end justify-between gap-4">
-        <div className="min-w-0">
-          <h2 className="text-xl font-bold text-slate-900 truncate">
-            {plc.name}
-          </h2>
-          <p className="text-sm text-slate-500 mt-0.5">
-            {t('plcDashboard.home.subtitle', {
-              defaultValue: 'Your collaborative space',
-            })}
-          </p>
+      {/* Page header — title/subtitle + members, then a quick-create bar */}
+      <div className="px-6 pt-6 pb-4 border-b border-slate-100 bg-white/50 backdrop-blur-sm">
+        <div className="flex items-end justify-between gap-4">
+          <div className="min-w-0">
+            <h2 className="text-xl font-bold text-slate-900 truncate">
+              {plc.name}
+            </h2>
+            <p className="text-sm text-slate-500 mt-0.5">
+              {t('plcDashboard.home.subtitle', {
+                defaultValue: 'Your collaborative space',
+              })}
+            </p>
+          </div>
+          <MembersHeaderCluster plc={plc} onNavigate={onNavigate} />
         </div>
-        <MembersHeaderCluster plc={plc} onNavigate={onNavigate} />
+        <div className="mt-4">
+          <QuickCreateBar onNavigate={onNavigate} />
+        </div>
       </div>
 
       {/* Two-column dashboard — fills the full content width */}
       <div className="flex-1 p-6">
         <div className="w-full grid grid-cols-1 lg:grid-cols-3 gap-5 items-start">
           {/* Primary column — what the PLC is working on */}
-          <div className="lg:col-span-2 flex flex-col gap-5">
+          <div className="lg:col-span-2">
             <AttentionCard plc={plc} onNavigate={onNavigate} />
-            <RecentDocsCard plc={plc} onNavigate={onNavigate} />
           </div>
 
-          {/* Sidebar column — quick actions */}
-          <div className="lg:col-span-1 flex flex-col gap-5">
-            <QuickCreateCard onNavigate={onNavigate} />
+          {/* Sidebar column — recent docs */}
+          <div className="lg:col-span-1">
+            <RecentDocsCard plc={plc} onNavigate={onNavigate} />
           </div>
         </div>
       </div>
